rearranging the video.py structure
